requirements.txt
streamlit
snowflake-ml-python
snowflake-snowpark-python